user registration page finished lesson 205 complete

+++ admin/includes/view_all_posts.php
@@ -46,7 +46,7 @@ if(isset($_POST['checkBoxArray'])){
         </div>
         <div class="col-xs-4">
             <input type="submit" name="submit" class="btn btn-success" value="Apply">
-            <a class="btn btn-primary" href="add_post.php">Add New</a>
+            <a class="btn btn-primary" href="posts.php?source=add_post">Add New</a>
         </div>
     
     <thead>
@@ -61,6 +61,7 @@ if(isset($_POST['checkBoxArray'])){
             <th>Tags</th>
             <th>Comments</th>
             <th>Date</th>
+            <th>View Post</th>
             <th>Edit</th>
             <th>Delete</th>
         </tr>
@@ -110,8 +111,9 @@ if(isset($_POST['checkBoxArray'])){
             echo "<td>$post_tags</td>";
             echo "<td>$post_comment_count</td>";
             echo "<td>$post_date</td>";
+            echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";    
             echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";    
-            echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";                                            
+            echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete'); \"href='posts.php?delete={$post_id}'>Delete</a></td>";                                            
             echo "</tr>";
                                         
         }
